requirements.txt
Django==1.11.5
pytz==2017.2
djangorestframework==3.7.3
